move all file to directory com.hzgc.dubbo.dynamic,
