+++ Dubbo/src/main/java/com/hzgc/dubbo/dynamic/PictureType.java
@@ -1,4 +1,4 @@
-package com.hzgc.dubbo.dynamic.picture;
+package com.hzgc.dubbo.dynamic;
 
 /**
  * 图片类型
