+++ Dubbo/src/main/java/com/hzgc/dubbo/dynamic/SearchType.java
@@ -1,4 +1,4 @@
-package com.hzgc.dubbo.dynamic.search;
+package com.hzgc.dubbo.dynamic;
 
 /**
  * 搜索类型，搜人或者搜车
